@@ -5,7 +5,6 @@ using Products.Memory;
 
 var builder = WebApplication.CreateBuilder(args);
 
-// add aspire service defaults
 builder.AddServiceDefaults();
 
 builder.Services.AddDbContext<ProductDataContext>(options =>
@@ -17,16 +16,19 @@ builder.Services.AddSingleton<IConfiguration>(sp =>
 });
 
 // add memory context
-builder.Services.AddSingleton<MemoryContext>(sp =>
+builder.AddSemanticKernel();
+
+if (builder.Environment.IsDevelopment())
 {
-    return new MemoryContext();
-});
+    builder.AddOllama();
+}
+else
+{
+    builder.AddAzureAI();
+}
 
 // Add services to the container.
 var app = builder.Build();
-
-// aspire map default endpoints
-app.MapDefaultEndpoints();
 
 // Configure the HTTP request pipeline.
 app.UseHttpsRedirection();
@@ -38,6 +40,8 @@ app.UseStaticFiles();
 app.CreateDbIfNotExists();
 
 // init semantic memory
-app.InitSemanticMemory();
+await app.InitSemanticMemoryAsync();
+
+app.MapDefaultEndpoints();
 
 app.Run();
