@@ -3,7 +3,7 @@
 @inject Store.Services.ProductService ProductService
 @inject IConfiguration Configuration
 @attribute [StreamRendering(true)]
-@rendermode InteractiveServer 
+@rendermode InteractiveServer
 
 <PageTitle>Search Products</PageTitle>
 
@@ -12,10 +12,14 @@
 <p>Search our amazing outdoor products that you can purchase.</p>
 
 <div class="form-group">
-    <label for="search">Type your question:</label>
-    <div class="input-group" >
-        <input css type="text" id="search" class="form-control" @bind="searchTerm" />
+    <label for="search" class="form-label">Type your question:</label>
+    <div class="input-group mb-3">
+        <input type="text" id="search" class="form-control" @bind="searchTerm" placeholder="Enter search term..." />
         <button id="btnSearch" class="btn btn-primary" @onclick="DoSearch" type="submit">Search</button>
+    </div>
+    <div class="form-check form-switch mb-3">
+        <InputCheckbox id="smartSearch2" @bind-Value="smartSearch" />
+        <label class="form-check-label" for="smartSearch">Use Semantic Search</label>
     </div>
     <hr />
     <p class="quote">@aiResponse</p>
@@ -23,6 +27,7 @@
     @* Uncomment this to show elapsed time *@
     @* <p class="smallquote">Elapsed Time: [@elapsedTime]</p> *@
 </div>
+
 
 @if (products == null)
 {
@@ -57,13 +62,14 @@ else
 @code {
     private string searchTerm = "";
     private List<DataEntities.Product>? products;
-    private string aiResponse = "";   
+    private string aiResponse = "";
     private string elapsedTime = "";
+    private bool smartSearch = false;
 
     private async Task DoSearch(MouseEventArgs e)
     {
         await Task.Delay(500);
-        var response = await ProductService.Search(searchTerm);
+        var response = await ProductService.Search(searchTerm, smartSearch);
         aiResponse = response.Response;
         products = response.Products;
         elapsedTime = response.ElapsedTime.ToString();
